feat: 북마크 지원자 상세 조회

+++ src/main/java/com/piveguyz/empickbackend/employment/applicant/query/dto/ApplicantBookmarkQueryDTO.java
@@ -7,8 +7,8 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ApplicantBookmarkQueryDTO {
-    private Long memberId;
-    private Long applicantId;
+    private int memberId;
+    private int applicantId;
     private String name;
     private String phone;
     private String email;
